refactor: rename cours complement => complement
