+++ src/core/data/convex/convex-impl.ts
@@ -13,7 +13,7 @@ import {
   Cours,
   Sequence,
 } from "@/features/cours-sequence/domain/entities/cours-schemas";
-import { CoursComplement } from "@/features/cours-complement/domain/cours-complement-schemas";
+import { CoursComplement } from "@/features/complement/domain/cours-complement-schemas";
 
 export interface ConvexDatabaseOptions {
   db: typeof api;
